refactor(file-manager): move safe open to file utils
