+++ src/renderer/components/chat/messages/hooks/useMessageLeafCapabilities.ts
@@ -3,11 +3,11 @@ import type { MessageListActions, MessageListState } from '@renderer/components/
 import { containsInlineFilePath } from '@renderer/components/chat/messages/utils/filePath'
 import { useAttachment } from '@renderer/hooks/useAttachment'
 import { useExternalApps } from '@renderer/hooks/useExternalApps'
-import { safeOpen } from '@renderer/services/safeOpen'
 import type { FileMetadata } from '@renderer/types/file'
 import type { McpTool } from '@renderer/types/tool'
 import { buildEditorUrl } from '@renderer/utils/editorUtils'
 import { parseFileTypes } from '@renderer/utils/file'
+import { safeOpen } from '@renderer/utils/file/safeOpen'
 import type { CherryMessagePart } from '@shared/data/types/message'
 import { IpcChannel } from '@shared/IpcChannel'
 import type { FileHandle, FilePath } from '@shared/types/file'
